Renamed existing method getProjectForDataset->getOrCreateProjectForDataset on the DataLocationService class to better describe the behavior. Added a new method called getProjectForDatasetId on the same class to look up existing projects only, not create new ones. Changed CreateDataset flight undoing direction to check if an associated project exists before attempting to call datasetService.retrieve, which tries to create it if it doesn't exist already.

+++ src/main/java/bio/terra/service/dataset/flight/create/CreateDatasetPrimaryDataStep.java
@@ -1,10 +1,12 @@
 package bio.terra.service.dataset.flight.create;
 
-import bio.terra.service.dataset.flight.DatasetWorkingMapKeys;
-import bio.terra.service.dataset.Dataset;
 import bio.terra.common.PrimaryDataAccess;
-import bio.terra.service.job.JobMapKeys;
+import bio.terra.service.dataset.Dataset;
+import bio.terra.service.dataset.DatasetDataProject;
 import bio.terra.service.dataset.DatasetService;
+import bio.terra.service.dataset.flight.DatasetWorkingMapKeys;
+import bio.terra.service.job.JobMapKeys;
+import bio.terra.service.resourcemanagement.DataLocationService;
 import bio.terra.stairway.FlightContext;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.Step;
@@ -16,10 +18,14 @@ import java.util.UUID;
 public class CreateDatasetPrimaryDataStep implements Step {
     private final PrimaryDataAccess pdao;
     private final DatasetService datasetService;
+    private final DataLocationService dataLocationService;
 
-    public CreateDatasetPrimaryDataStep(PrimaryDataAccess pdao, DatasetService datasetService) {
+    public CreateDatasetPrimaryDataStep(PrimaryDataAccess pdao,
+                                        DatasetService datasetService,
+                                        DataLocationService dataLocationService) {
         this.pdao = pdao;
         this.datasetService = datasetService;
+        this.dataLocationService = dataLocationService;
     }
 
     Dataset getDataset(FlightContext context) {
@@ -38,10 +44,15 @@ public class CreateDatasetPrimaryDataStep implements Step {
 
     @Override
     public StepResult undoStep(FlightContext context) {
-        try {
+        // check if project was created
+        UUID datasetId = context.getWorkingMap().get(DatasetWorkingMapKeys.DATASET_ID, UUID.class);
+        DatasetDataProject datasetProject = dataLocationService.getProjectForDatasetId(datasetId);
+        if (datasetProject != null) {
+            // if there is no project associated with this dataset, then there won't be primary data to delete,
+            // so no need to call the below method.
+            // also the below method will try to create a project if one doesn't already exist, and we don't
+            // want to do that in the undo direction.
             pdao.deleteDataset(getDataset(context));
-        } catch (Exception ex) {
-            System.out.println("exception caught 21. "+ex.getMessage()+", "+ex.getClass().getCanonicalName());
         }
         return StepResult.getStepResultSuccess();
     }
